+++ Service/main_infrared_camera.py
@@ -443,17 +443,18 @@ class Delete_file(MyQThread):
         total_nums = 0
         for root, dirs, files in os.walk(self.path):
             # logger.warning(f"{root} | {dirs} | {files}")
-            for file in files:
-                file_path = os.path.join(root, file)
-                try:
-                    size = os.path.getsize(file_path)  # 获取文件大小（字节）
-                    size = float(size / 1204 / 1024 / 1024)  # 将字节B转成GB
-                    total_size += size
-                    total_nums += 1
-                    os.remove(file_path)  # 删除文件
-                except Exception as e:
-                    logger.error(
-                        f"infrared_camera Failed to delete {file_path}: reason:{e} |  异常堆栈跟踪：{traceback.print_exc()}")
+            if "tmp" not in root:
+                for file in files:
+                    file_path = os.path.join(root, file)
+                    try:
+                        size = os.path.getsize(file_path)  # 获取文件大小（字节）
+                        size = float(size / 1204 / 1024 / 1024)  # 将字节B转成GB
+                        total_size += size
+                        total_nums += 1
+                        os.remove(file_path)  # 删除文件
+                    except Exception as e:
+                        logger.error(
+                            f"infrared_camera Failed to delete {file_path}: reason:{e} |  异常堆栈跟踪：{traceback.print_exc()}")
         global frame_nums
         with lock:
             logger.warning(f"红外相机 | 删除文件总大小: {total_size} G-bytes | 删除文件总数量： {total_nums} | 此时相机拍摄的图像数量：{frame_nums}")
